pyproject.toml
[project]
name = "sinamecc-backend"
version = "1.0.0"
description = "Backend for the Sinamecc project"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.35.0,<1.40.0",
    "django>=5.1.5",
    "django-cors-headers>=4.6.0,<4.7.0",
    "django-environ>=0.11.0,<0.12.0",
    "django-fsm==3.0.0",
    "django-role-permissions==3.2.0",
    "django-storages>=1.14,<1.15",
    "django-viewflow>=2.2.9,<2.6",
    "djangorestframework>=3.15,<3.16",
    "djangorestframework-simplejwt>=5.3.1,<5.4.0",
    "gunicorn>=23,<24",
    "psycopg2-binary>=2.9.0,<2.10.0",
]

[dependency-groups]
dev = [
    "moto>=5.0.0,<5.1.0",
    "pytest-django>=4.9.0,<5.0.0",
    "pytest-factoryboy>=2.7.0,<3.0.0",
    "pytest-mock>=3.14.0",
]
[tool.pytest]
addopts = ["-p", "no:warnings"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.local"
# -- recommended but optional:
python_files = ["test_*.py"]
pythonpath = "/"
markers = [
    "users: mark a test as a users test",
    "testfast: mark a test as a fast test",
]
